Bump [compat], fix PDB download format
The latest version of BioStructures downloads MMCIFF by default, which
breaks the later MIToS test. This specifies the format. Also bumps
[compat] on many dependencies.

Makie is excluded from the bump because the latest version removes
`AbstractPoint`. This is presumably easy to fix but it seems worth
getting a new release out that doesn't require any internal changes.

Closes #627
Closes #628
Closes #631
Closes #632
Closes #633
Closes #634
Closes #635
Closes #642
Closes #643