[compat] julia 1.10+, test on CI
Formerly this package declared compatibility with any Julia 1.x.y, but
only ran tests on "1" (meaning the current release). Of course, it would
be surprising if this package still worked in its current form on Julia
1.0.0. However, since we don't run tests for anything but the current
version of Julia, we have no way of knowing what it takes for it to
work, and consequently no way of ensuring that users with working
systems won't get an upgrade that breaks BioMakie for them.

Fortunately, the current LTS, Julia 1.10, is quite recent and I can
verify that the package works on that version. To put matters on sound
footing for future development, this PR adopts 1.10 as the minimum
version, and adds a test for Julia 1.10 to the CI workflow.